requirements.txt
# Core Telegram library
pyrogram[tgcrypto,Pillow]>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Numerical operations
numpy>=1.20.0

# Machine Learning (check PyTorch website for specific CUDA/CPU versions)
torch>=1.8.0
torchvision>=0.9.0

# Face detection and recognition
facenet-pytorch>=2.5.0

# Image processing
opencv-python>=4.5.0

# Progress bar for indexer
tqdm>=4.60.0 